feat: Update to WMS.Data.Contracts.WebAPI release 0.16.25

+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService.Contracts/ServiceContracts.autogen.cs
@@ -1323,11 +1323,11 @@ namespace Ferretto.VW.MAS.AutomationService.Contracts
         System.Threading.Tasks.Task<LoadingUnitDetails> GetWmsDetailsByIdAsync(int id, System.Threading.CancellationToken cancellationToken);
     
         /// <exception cref="MasWebApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task ImmediateAddItemAsync(int id, int itemId, int quantity, int compartmentId, string lot, string serialNumber);
+        System.Threading.Tasks.Task ImmediateAddItemAsync(int id, int itemId, double quantity, int compartmentId, string lot, string serialNumber);
     
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <exception cref="MasWebApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task ImmediateAddItemAsync(int id, int itemId, int quantity, int compartmentId, string lot, string serialNumber, System.Threading.CancellationToken cancellationToken);
+        System.Threading.Tasks.Task ImmediateAddItemAsync(int id, int itemId, double quantity, int compartmentId, string lot, string serialNumber, System.Threading.CancellationToken cancellationToken);
     
         /// <exception cref="MasWebApiException">A server side error occurred.</exception>
         System.Threading.Tasks.Task InsertLoadingUnitAsync(LoadingUnitLocation source, int? destinationCellId, int loadingUnitId);
