+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Controllers/LoadingUnitsController.cs
@@ -259,7 +259,7 @@ namespace Ferretto.VW.MAS.AutomationService.Controllers
         [HttpPost("{id}/immediate-additem")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<IActionResult> ImmediateAddItemAsync(int id, int itemId, int quantity, int compartmentId, string lot, string serialNumber, [FromServices] ILoadingUnitsWmsWebService loadingUnitsWmsWebService)
+        public async Task<IActionResult> ImmediateAddItemAsync(int id, int itemId, double quantity, int compartmentId, string lot, string serialNumber, [FromServices] ILoadingUnitsWmsWebService loadingUnitsWmsWebService)
         {
             //public async Task<IActionResult> ImmediateAddItemAsync(int id, int itemId, int quantity, int compartmentId, [FromServices] ILoadingUnitsWmsWebService loadingUnitsWmsWebService)
             if (loadingUnitsWmsWebService is null)
